Test different data types

+++ test/onnx/gen_onnx.py
@@ -333,12 +333,12 @@ def batch_norm_flat_test():
 
 @onnx_test
 def batch_norm_1d_test():
-    x = helper.make_tensor_value_info('x', TensorProto.FLOAT, [2, 3, 4])
+    x = helper.make_tensor_value_info('x', TensorProto.FLOAT16, [2, 3, 4])
     scale = helper.make_tensor_value_info('scale', TensorProto.FLOAT, [3])
     bias = helper.make_tensor_value_info('bias', TensorProto.FLOAT, [3])
     mean = helper.make_tensor_value_info('mean', TensorProto.FLOAT, [3])
     var = helper.make_tensor_value_info('variance', TensorProto.FLOAT, [3])
-    out = helper.make_tensor_value_info('y', TensorProto.FLOAT, [2, 3, 4])
+    out = helper.make_tensor_value_info('y', TensorProto.FLOAT16, [2, 3, 4])
 
     node = onnx.helper.make_node(
         'BatchNormalization',
@@ -368,12 +368,12 @@ def batch_norm_2d_test():
 
 @onnx_test
 def batch_norm_3d_test():
-    x = helper.make_tensor_value_info('x', TensorProto.FLOAT, [2, 2, 2, 2, 2])
-    scale = helper.make_tensor_value_info('scale', TensorProto.FLOAT, [2])
-    bias = helper.make_tensor_value_info('bias', TensorProto.FLOAT, [2])
-    mean = helper.make_tensor_value_info('mean', TensorProto.FLOAT, [2])
-    var = helper.make_tensor_value_info('variance', TensorProto.FLOAT, [2])
-    out = helper.make_tensor_value_info('y', TensorProto.FLOAT,
+    x = helper.make_tensor_value_info('x', TensorProto.DOUBLE, [2, 2, 2, 2, 2])
+    scale = helper.make_tensor_value_info('scale', TensorProto.DOUBLE, [2])
+    bias = helper.make_tensor_value_info('bias', TensorProto.DOUBLE, [2])
+    mean = helper.make_tensor_value_info('mean', TensorProto.DOUBLE, [2])
+    var = helper.make_tensor_value_info('variance', TensorProto.DOUBLE, [2])
+    out = helper.make_tensor_value_info('y', TensorProto.DOUBLE,
                                         [2, 2, 2, 2, 2])
 
     node = onnx.helper.make_node(
